Bug Fix: A defaultBankIndex and defaultPresetIndex are now set in synth.setSoundFont(...)...
The default, in each channel, is the first preset in the first WAF to be defined.
N.B.: In the ResidentWAFSynthHost application (in the current repository), the bank and preset defaults are overridden when the host initializes so that each channel is given its own preset. The lower level default is, however, useful for other hosts.